--- a/requirements.txt
+++ b/requirements.txt
@@ -1,3 +1,4 @@
 google-auth-oauthlib>=1.0.0
 requests>=2.25.0
 python-dotenv>=0.19.0
+pyyaml>=6.0.0